--- a/daemon/graphdriver/overlay2/overlay_unsupported.go
+++ b/daemon/graphdriver/overlay2/overlay_unsupported.go
@@ -1,0 +1,3 @@
+// +build !linux
+
+package overlay2