@dmcgowan 2016-04-18
One additional tweak to the design I am making now is to always have a `lower-id` and exclude the root from the `lower`. This will ensure that the combination of the `diff` and `lowers` is equivalent to the prior `upper` directory. It should allow cleaning up some logic and ease some of the cleanup when there is a squash failure.