@unclejack 2016-04-22
@dmcgowan Does this fix some other issues we were encountering when using Overlay?