@dmcgowan 2016-04-23
Rebased to account for #22168. Might be worth discussing how this will impact this PR. The fsdiff is still used in some cases but not all.