@cpuguy83 2016-05-10
Pretty awesome!
Verified this works and seems to work well. I have not done any performance comparisons or stress testing, but functionally it seems pretty solid.

The only problem with the current approach is that it's not backwards compatible.
So a user using overlay can upgrade Docker get this changed out under the hood without noticing any changes, but then if they need to downgrade anything new will not work under the old driver.

It seems like this either needs to be a new driver (`overlay2`?) so there are no surprises here, or the newer version needs to setup new layers in a backwards compatible way.