@dmcgowan 2016-05-10
@tonistiigi: just some of the initialization checks. That is fair to point out, the duplication is because the current approach can be used on top of an existing overlay directory and uses the current hard link method to "squash" layers past the options limit into upper directories as used by the current driver.

I found that squashing is required when an image gets about 35 layers deep when using the default "/var/lib/docker" location.