@dmcgowan 2016-05-10
@unclejack not late at all. Did you have any specific issues in mind? I am going back through the issue list to see which are reproducible and not identified as Kernel bugs.

These are the issues I know are addressed, would love to add to the list
- #10613 - Inode exhaustion, fixed by using less files
- #17551 - Slow delete, fixed by having less files to remove
- #21555 - Unstable images, fixed by using a diff directory