@unclejack 2016-05-10
@dmcgowan https://github.com/docker/docker/issues/9572 https://github.com/docker/docker/issues/19647 https://github.com/docker/docker/issues/12080

These are not necessarily issues anyone would expect this PR to fix, just something I'd hoped this change would fix through multiple lower directories.