@crosbymichael 2016-05-12
@dmcgowan does overlay work with `MS_REMOUNT` for specifying more layers after it is already mounted or do we have no choice but to use hardlinks to squash?