@crosbymichael 2016-05-13
Would something like this work to get away from all of the hard linking or would the multiple mounts be worse?

```go
package main

import (
	"fmt"
	"os"
	"path/filepath"
	"syscall"

	"github.com/Sirupsen/logrus"
)

func main() {
	if err := run(); err != nil {
		logrus.Fatal(err)
	}
}

func run() error {
	cwd, err := os.Getwd()
	if err != nil {
		return err
	}
	for _, name := range []string{
		"lower1",
		"lower2",
		"merged",
		"upper",
		"work",
	} {
		if err := os.MkdirAll(filepath.Join(cwd, name), 0755); err != nil {
			return err
		}
	}
	// mount the RO layers first so we don't overload the mount syscall with too much data
	if err := syscall.Mount(
		"overlay",
		filepath.Join(cwd, "merged"),
		"overlay",
		0,
		fmt.Sprintf(
			"lowerdir=%s:%s,workdir=%s",
			filepath.Join(cwd, "lower1"),
			filepath.Join(cwd, "lower2"),
			filepath.Join(cwd, "work"))); err != nil {
		return err
	}
	// now mount the rw layer with the already mounted lower's merge dir as the lower dir here
	if err := syscall.Mount(
		"overlay",
		filepath.Join(cwd, "merged"),
		"overlay",
		0,
		fmt.Sprintf(
			"lowerdir=%s,workdir=%s,upperdir=%s",
			filepath.Join(cwd, "merged"),
			filepath.Join(cwd, "work"),
			filepath.Join(cwd, "upper"))); err != nil {
		return err
	}
	return nil
}
```