@tonistiigi 2016-05-13
@crosbymichael  @vikstrous tried something like that in https://github.com/docker/docker/pull/18560#issuecomment-165947583 but hit some blockers.

There are other ways to get over the options limit. For example symlinks and relative paths.