@crosbymichael 2016-05-13
I think my preference would be to make this change a new driver.  It would be nice to have the new "pure" implementation and not have it messed up by the old one.  Then as time passes we can eventually fade out the old implementation.  

I'm fine whipping out my var/lib/docker for the new implementation as its much better and cleaner.   This is just my initial thought and I don't feel too strongly about same or different drivers, whatever is cleaner and easier to maintain is better for the long run.