@dmcgowan 2016-05-13
@crosbymichael @tonistiigi: going to make a proof of concept using the symlink and relative path method. I calculated that we should be able to hit a max depth of 128, which should allow making the cleaner separate driver.

The approach I plan on taking is generating a random 26 character identifier for each layer and creating a symlink named with that identifier in a link directory pointing to the diff directory. On mount the process with fork, change directories and use the relative path to the links for the lower dirs.