@dmcgowan 2016-05-13
Added the separate driver with no squashing and support for 128 layer depth. I will remove the other implementation from this PR once I have completed testing and benchmarking. I will update the PR description accordingly as well. The initial version is calling out to "/usr/bin/mount" but updating to use `reexec` instead. This is also still using the "<id>/diff" ordering instead of the "diff/<id>" order aufs uses. I don't have much of an opinion on this but don't mind updating if the consensus is to use the aufs method.