@dmcgowan 2016-05-17
Updated PR description for the separate driver. Also re-ran the usage tests and included the `wordpress` image. It is an interesting case since it has many layers but is not a very large image.