@dmcgowan 2016-05-24
Updated to use directory based reference counter. Also added a unit test for changes and fixed a bug which it exposed.